Move filepath linting to separate file
